+++ Recipe.rb
@@ -39,39 +39,7 @@ class Recipe
   end
 end
 
-class RecipeAct
-  def self.id_from_user
-    puts ''
-    puts 'Puts Needed Recipe ID:'
-    recipe_id_input = gets.chomp
-    with_recipe(recipe_id_input)
-  end
 
-  def self.with_recipe(recipe_id)
-    recipe_id = recipe_id.to_i
-    DATA.all_recipes[recipe_id - 1].show_recipe
-    while true
-      puts '1. Add to favorites'
-      puts '2. Delete recipe'
-      puts '0. Back to main'
-      puts ''
-      puts 'Enter a number '
-      user_input = gets.chomp
-      case user_input
-      when '1'
-        DATA.all_recipes[recipe_id - 1].is_favorite_change
-        DATA.save_halper
-        puts('Added to favorites')
-      when '2'
-        DATA.delete_from_data(recipe_id)
-        puts('Delete Recipe')
-      when '0'
-        DATA.save_halper
-        Menu_main.general_main
-      end
-    end
-  end
-end
 
 class RecipeStore
   attr_accessor :recipe_list_hash, :recipe_list_obj
